fix: add font vertical offset
Add an offset to the font using `offset.y` to add some vertical spacing
between each line to make things easier to read and more like I had it
in PhpStorm.

Also use `glyph.offset` to ensure that the spacing is balanced between
the lines rather than all being above each line, as this looks a bit odd
in Vim and with the tmux status line.