chore(nvim): disable auto-formatting
This is tricky when working on projects that use different formatting
and coding styles (e.g. Drupal and Symfony), and on projects that
haven't had code formatting run on them before and making a small change
causes a large number of changes to be staged.

I'll experiment with formatting code within a range, or toggling
auto-formatting per project. Maybe something similar to what @wincent
showed in https://youtu.be/ifA8iw42vwI.