feat(nvim): add drupalclass snippet
Add a new `drupalclass` snippet for Luasnip that scaffolds a new PHP
class that's opinionated for Drupal applications.

The class name is based on the filename (minus the file extension) and
the namespace is generated automatically from the directory structure
based on the location of the `src` directory.

It also includes the `Tests` entry within the namespace if the file is
witin a `tests` directory.