feat(tmux): add `join-pane` key bind
Add a key binding for re-joining another window into a pane within the
current window, which is the opposite of the `break-pane` command.

It prompts for the window to join from and then joins it into a vertical
pane within the current window.

This was taken from Chris Toomey's dotfiles. See
https://github.com/christoomey/dotfiles/blob/553c92a3a1fd7f05f2283ab8c45feffcefea5d5b/tmux/tmux.conf#L60