fix(scripts): vendor files causing reloads
Ignore the `vendor` directory and some others as it was causing the
script to reload when it wasn't expected.

Whilst I'd expect it to be ignored if I specified the path to look at as
something like `web/modules/custom`, but it still seems to trigger the
command and it seems a lot more stable since adding this.