feat(nvim): use rstfmt for rst file formatting
Add `vimPlugins.conform-nvim` to add formatting rst files with `rstfmt`.

This isn't supported as a formatter in `none-ls`, and I wanted to try
and potentially switch to `conform-nvim` and `nvim-lint` in the future,
so this is a potential first step towards that.

`rstfmt` itself is already installed and available within my additional
Vim packages.

This currently needs to be run manually, as I don't want to change my
default keybind for formatting until I've supported all the filetypes and formatters
I have in `none-ls`.

To run the foramt manually, run `:lua require"conform".format()`.

See:

- https://github.com/stevearc/conform.nvim
- https://github.com/mfussenegger/nvim-lint
- https://pypi.org/project/rstfmt
- https://www.youtube.com/watch?v=ybUE4D80XSk