Add `todo` and `til` functions
Add helper functions to write text input into a `TODO.txt` or `TIL.txt`
file respectively.

Extracting this to a function means I'm not able to overwrite the
contents of a file by typing `echo "foo" > TODO.txt` or similar
accidentally and overwriting the entire file instead of appending to it.