Update `t` script
Stop automatically running `.tmux` files as I don't always want
everything to run when opening a project. This also simplifies the code
within the script.

This also refactors the logic for creating and switching sessions as it
didn't work 100% of the time. This is nearer to Jess Archer's version of
the script with less error output and changes needed for shellcheck.